requirements.txt
babel==2.17.0
bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
Flask==3.1.2
flask-babel==4.0.0
flask-mail==0.10.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pytz==2025.2
Werkzeug==3.1.3
Pillow==11.1.0
requests==2.32.3
python-dotenv==1.0.1
